pyproject.toml
[tool.poetry]
name = "simvue"
version = "1.0.1"
description = "Simulation tracking and monitoring"
authors = ["Simvue Development Team <info@simvue.io>"]
license = "Apache v2"
readme = "README.md"
homepage = "https://simvue.io"
repository = "https://github.com/simvue-io/python-api"
documentation = "https://docs.simvue.io"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: Unix",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
    "Typing :: Typed"
]

keywords = [
    "tracking",
    "monitoring",
    "metrics",
    "alerting",
    "metrics-gathering"
]

[tool.poetry.dependencies]
python = "^3.9"
dill = "^0.3.7"
requests = "^2.31.0"
msgpack = "^1.0.7"
tenacity = ">=8.2.3,<10.0.0"
PyJWT = "^2.8.0"
psutil = "^5.9.8"
pydantic = "^2.5.3"
pandas = "^2.2.0"
plotly = {version = "^5.18.0", optional = true}
numpy = ">=1.26.3,<3.0.0"
matplotlib = {version = "^3.8.2", optional = true}
typing_extensions = { version = "^4.11.0", python = "<3.10" }
toml = "^0.10.2"
click = "^8.1.7"
gitpython = "^3.1.43"
humanfriendly = "^10.0"
tabulate = "^0.9.0"
randomname = "^0.2.1"

[tool.poetry.extras]
plot = ["matplotlib", "plotly"]

[tool.poetry.scripts]
simvue_sender = "simvue.bin.sender:run"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-dependency = "^0.6.0"
ruff = ">=0.5.0,<0.6.0"
pytest-cov = ">=4.1,<6.0"
pytest-mock = "^3.14.0"
pytest-sugar = "^1.0.0"
pytest-xdist = "^3.6.1"
jinja2 = "^3.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[tool.ruff]
lint.extend-select = ["C901", "T201"]
lint.mccabe.max-complexity = 11

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T201"]
"examples/*" = ["T201"]
"notebooks/*" = ["T201"]

[tool.pytest.ini_options]
addopts = "-p no:warnings --no-cov -n 0"
testpaths = [
    "tests"
]
markers = [
    "client: tests of Simvue client",
    "converters: tests for Simvue object converters",
    "dispatch: test data dispatcher",
    "run: test the simvue Run class",
    "utilities: test simvue utilities module",
    "scenario: test scenarios",
    "executor: tests of executors",
    "config: tests of simvue configuration",
    "api: tests of RestAPI functionality"
]

[tool.interrogate]
ignore-init-method = true
fail-under = 95
verbose = 1
exclude = ["docs", "tests"]
